fix service of order and model [error -> int - long]

+++ src/main/java/blc/idat/apiblc/rest/PedidoController.java
@@ -16,7 +16,7 @@ public class PedidoController {
     private PedidoService pedService;
 
     @GetMapping("/all/{id}")
-    public List<Pedido> findByCliente(@PathVariable("id") int id){
+    public List<Pedido> findByCliente(@PathVariable("id") Long id){
         return pedService.findbyCliente(id);
     }
 }
